requirements.txt
asgiref==3.5.2
asttokens==2.0.8
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.1
commonmark==0.9.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.1
django-filter==22.1
djangorestframework==3.13.1
executing==1.0.0
flake8==5.0.4
flake8-builtins==1.5.3
flake8-django==1.1.5
flake8-import-order==0.18.1
flake8-print==5.0.0
httpie==3.2.1
idna==3.3
importlib-metadata==4.12.0
ipython==8.5.0
jedi==0.18.1
Markdown==3.4.1
matplotlib-inline==0.1.6
mccabe==0.7.0
multidict==6.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.13.0
PySocks==1.7.1
pytz==2022.2.1
requests==2.28.1
requests-toolbelt==0.9.1
rich==12.5.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.5.0
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
zipp==3.8.1
